Consolidate state methods into state_helpers
