+++ ui/bookmarks/reducers.py
@@ -1,12 +1,12 @@
 from frozendict import FrozenOrderedDict
 from collections import OrderedDict
 from ..book.reducers import BookReducers
-from .. import utility
+from .. import state_helpers
 
 
 class BookmarksReducers():
     def delete_bookmark(self, state, n):
-        width, height = utility.dimensions(state)
+        width, height = state_helpers.dimensions(state)
         # adjust for title
         height -= 1
         page = state['bookmarks_menu']['page']
@@ -28,7 +28,7 @@ class BookmarksReducers():
         return state.copy(user=state['user'].copy(books=FrozenOrderedDict(books)))
 
     def go_to_bookmark(self, state, n):
-        width, height = utility.dimensions(state)
+        width, height = state_helpers.dimensions(state)
         # adjust for title
         height -= 1
         page = state['bookmarks_menu']['page']
