+++ ui/library/reducers.py
@@ -2,19 +2,18 @@ from frozendict import FrozenOrderedDict
 import logging
 from collections import OrderedDict
 
-from .get_books import get_books
 from ..book import book_file
-from .. import utility
+from .. import state_helpers
 
 log = logging.getLogger(__name__)
 
 
 class LibraryReducers():
     def go_to_book(self, state, number):
-        width, height = utility.dimensions(state)
+        width, height = state_helpers.dimensions(state)
         page = state['library']['page']
         line_number = page * (height - 1)
-        books = get_books(state)
+        books = state_helpers.get_books(state)
         try:
             book = books[line_number + number]
         except:
