+++ ui/library/view.py
@@ -1,5 +1,5 @@
 from ..braille import from_ascii, format_title
-from .get_books import get_books
+from ..state_helpers import get_books
 
 
 def render_help_menu(width, height, page):
